added custom vectors
