+++ _includes/code/quickstart.byov.schema.mdx
@@ -31,12 +31,25 @@ import ByovAllPy3Code from '!!raw-loader!/_includes/code/quickstart.byov.all_v3.
 
   </TabItem>
 
-  <TabItem value="js" label="JS/TS (Legacy)">
+  <TabItem value="js" label="JS/TS">
 
 import ByovAllTsCode from '!!raw-loader!/_includes/code/quickstart.byov.all.ts';
 
 <FilteredTextBlock
   text={ByovAllTsCode}
+  startMarker="// Class definition"
+  endMarker="// Import data"
+  language="ts"
+/>
+
+  </TabItem>
+
+  <TabItem value="js2" label="JS/TS (Legacy)">
+
+import ByovAllTsCodeLegacy from '!!raw-loader!/_includes/code/quickstart.byov.all.legacy.ts';
+
+<FilteredTextBlock
+  text={ByovAllTsCodeLegacy}
   startMarker="// Class definition"
   endMarker="// Import data"
   language="ts"
