+++ _includes/code/quickstart/clients.install.mdx
@@ -24,7 +24,16 @@ pip install "weaviate-client==3.*"
 ```
 
 </TabItem>
-<TabItem value="js" label="JavaScript/TypeScript">
+<TabItem value="js" label="JS/TS">
+
+Add `weaviate-client` to your project with `npm`:<br/><br/>
+
+```bash
+npm install weaviate-client
+```
+
+</TabItem>
+<TabItem value="js2" label="JS/TS (Legacy)">
 
 Add `weaviate-ts-client` to your project with `npm`:<br/><br/>
