+++ package.json
@@ -68,7 +68,8 @@
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6",
     "weaviate-ts-client": "^2.1.0",
-    "weaviate-ts-embedded": "^1.2.0"
+    "weaviate-ts-embedded": "^1.2.0",
+    "weaviate-client": "^3.0.0-beta.8"
   },
   "browserslist": {
     "production": [
